requirements.txt
# requirements.txt
crewai
crewai-tools
streamlit
sqlalchemy
requests
beautifulsoup4
python-dotenv
openai